moved internally used methods to ./src/utils/
